Refactor: Decouple Freeze Voting from Core Strategy via Adapters
**Context & Problem:**

The previous freeze voting mechanism was tightly coupled with specific token types (ERC20/ERC721 Freeze Voting contracts). This design had limitations:
1.  It required deploying a different type of freeze voting contract depending on the parent DAO's primary governance token type.
2.  It didn't easily support DAOs with multiple voting token types (e.g., both ERC20 and ERC721 for voting power) for participating in freeze votes.
3.  Extending freeze voting to support new types of voting power (e.g., based on other criteria like Hats, or future token standards) would require new, separate freeze voting contracts, leading to code duplication and complexity.

The goal of this work was to refactor the freeze voting system to be more flexible, extensible, and aligned with the adapter pattern already used for normal proposal voting power and proposer eligibility in the `StrategyV1` contract.

**Summary of Changes:**

This refactor introduces a more dynamic and unified approach to freeze voting, primarily by:

1.  **New `IFreezeVoteContributorV1` Interface:**
    *   Defines a standard interface (`getFreezeVoteWeight`, `recordFreezeVote`) that any voting adapter can implement to contribute its specific voting power calculation and duplicate-vote prevention logic to a freeze vote.
    *   Includes a `FreezeVoteRecorded` event and common errors (`UnauthorizedFreezeVoter`, `NoFreezeVotingWeight`).

2.  **Unified `AzoriusFreezeVotingV1` Contract:**
    *   Replaces the previous `ERC20FreezeVotingV1` and `ERC721FreezeVotingV1` contracts.
    *   This single contract orchestrates freeze votes by interacting with the parent DAO's `StrategyV1` to identify its active voting adapters.
    *   It then calls `recordFreezeVote` on any of these adapters that implement `IFreezeVoteContributorV1`, allowing diverse sources of voting power to contribute to a freeze decision.
    *   Manages the freeze proposal lifecycle (creation timestamp, vote counting, periods) by inheriting from `BaseFreezeVotingV1`.
    *   The `FreezeProposalCreated` event (now in `IAzoriusFreezeVotingV1`) now includes the `strategy` address used for the freeze vote.

3.  **Voting Adapter Modifications (`ERC20VotingAdapterV1`, `ERC721VotingAdapterV1`):**
    *   Both adapters now implement `IFreezeVoteContributorV1`.
    *   They include new internal state to track freeze votes separately from normal proposal votes, namespaced by the calling `AzoriusFreezeVotingV1` contract (representing the child DAO) and the `freezeProposalSnapshotAndId`. This ensures votes for different child DAOs or different freeze proposal sessions do not interfere.
    *   The `recordFreezeVote` functions are protected by a new `onlyAuthorizedFreezeVoter` modifier.

4.  **`BaseVotingAdapterV1` Enhancement:**
    *   The `onlyAuthorizedFreezeVoter` modifier has been added here for reusability by all voting adapters. This modifier checks if the `msg.sender` (the `AzoriusFreezeVotingV1` contract) is authorized by the parent's `StrategyV1`.

5.  **`StrategyV1` and `IStrategyV1` Enhancements:**
    *   The role previously known as `_proposalInitializer` (and its accessor `proposalInitializer()`, modifier `onlyProposalInitializer`) has been renamed to `_strategyAdmin` (accessor `strategyAdmin()`, modifier `onlyStrategyAdmin`) to better reflect its broader administrative responsibilities, including managing freeze voter authorizations. A new error `InvalidStrategyAdmin` was added.
    *   New state variables (`_authorizedFreezeVotersMapping`, `_authorizedFreezeVotersArray`) and functions (`addAuthorizedFreezeVoter`, `removeAuthorizedFreezeVoter`, `isAuthorizedFreezeVoter`, `authorizedFreezeVoters`) were added to allow the `StrategyV1` (controlled by its `_strategyAdmin`) to maintain a list of authorized `AzoriusFreezeVotingV1` contracts that are permitted to trigger freeze vote recording on its adapters.
    *   A `FreezeVoterAuthorizationChanged` event is emitted when this list is modified.

6.  **`AzoriusFreezeGuardV1` Update:**
    *   Confirmed to use `IBaseFreezeVotingV1` for its `_freezeVoting` variable, ensuring compatibility with any contract implementing this base interface (including the old `MultisigFreezeVotingV1` and the new `AzoriusFreezeVotingV1`).

7.  **Timestamp Handling for Freeze Votes:**
    *   The `AzoriusFreezeVotingV1` contract now sets `_freezeProposalCreated = uint48(block.timestamp - 1)` when a freeze proposal period starts. This value serves as both the unique identifier for the freeze proposal session and the snapshot timestamp passed to voting adapters.
    *   The `IFreezeVoteContributorV1` interface functions (`getFreezeVoteWeight`, `recordFreezeVote`) now use a single `freezeProposalSnapshotAndId` parameter.

8.  **Mock Contracts (`MockVotingStrategy`, `MockAzoriusV1`, `MockFreezeVoteContributorAdapter`, `ConcreteBaseVotingAdapterV1`):**
    *   Updated or created to support the new interfaces and functionalities, enabling comprehensive unit testing.

9.  **Unit Tests:**
    *   Added extensive unit tests for the new freeze voting functionality in `ERC20VotingAdapterV1.test.ts` and `ERC721VotingAdapterV1.test.ts`, covering `getFreezeVoteWeight`, `recordFreezeVote` (including authorization and duplicate vote scenarios).
    *   Added tests for the `onlyAuthorizedFreezeVoter` modifier in `BaseVotingAdapterV1.test.ts`.
    *   Added tests for the new authorization management functions in `StrategyV1.test.ts`.
    *   Set up the initial test structure for `AzoriusFreezeVotingV1.test.ts`.
    *   Resolved various test failures and linter errors throughout the process.

**Benefits of these Changes:**

*   **Flexibility:** Freeze votes can now aggregate voting power from any combination of voting adapters configured on the parent DAO's strategy, provided those adapters implement `IFreezeVoteContributorV1`.
*   **Extensibility:** Supporting new types of voting power for freeze votes now only requires the new adapter to implement the standard interface. No new freeze voting orchestrator contract is needed.
*   **Code Consolidation:** Replaced two specialized freeze voting contracts with a single, more generic `AzoriusFreezeVotingV1`.
*   **Clearer Roles:** Renamed `proposalInitializer` to `strategyAdmin` for better semantic accuracy.
*   **Improved Testability:** Enhanced mock contracts and added comprehensive tests for the new mechanisms.

This refactor significantly improves the architecture of the freeze voting system, making it more robust, maintainable, and future-proof.

+++ contracts/interfaces/decent/deployables/IAzoriusV1.sol
@@ -6,7 +6,7 @@ import {Transaction} from "../Module.sol";
 interface IAzoriusV1 {
     event ProposalCreated(
         address strategy,
-        uint256 proposalId,
+        uint32 proposalId,
         address proposer,
         Transaction[] transactions,
         string metadata
